rename node.h to ir_base.h
