+++ cinn/common/cinn_value.cc
@@ -1,7 +1,7 @@
 #include "cinn/common/cinn_value.h"
 
 #include "cinn/ir/ir.h"
-#include "cinn/ir/node.h"
+#include "cinn/ir/ir_base.h"
 #include "cinn/poly/stage.h"
 #include "cinn/runtime/cinn_runtime.h"
