+++ cinn/hlir/pe/transform.h
@@ -2,7 +2,7 @@
 #include <string>
 #include <vector>
 
-#include "cinn/ir/node.h"
+#include "cinn/ir/ir_base.h"
 
 namespace cinn {
 namespace hlir {
