+++ cinn/ir/lowered_func.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 #include "cinn/ir/buffer.h"
-#include "cinn/ir/node.h"
+#include "cinn/ir/ir_base.h"
 
 namespace cinn {
 namespace ir {
